implementato commenti, aggiunto css per form

+++ siw-movie-00/src/main/java/it/uniroma3/siw/upload/FileUploadUtil.java
@@ -16,15 +16,21 @@ import org.springframework.web.multipart.MultipartFile;
 public class FileUploadUtil {
 
 	public static void saveFile(String uploadDir, String fileName, MultipartFile multipartFile) throws IOException {
-		
-		while (Files.exists(Paths.get(uploadDir)))								// se il path era usato precedentemente cancellalo + il vecchio file
-			FileUtils.deleteDirectory(new File(uploadDir));	
+
+		if (Files.exists(Paths.get(uploadDir))) { 			// se il path era usato precedentemente cancellalo + il vecchio file
+			FileUtils.deleteDirectory(new File(uploadDir));
+		}
+
+		// while (Files.exists(Paths.get(uploadDir)))
+		// 		System.out.println("deleting " + uploadDir + " folder...");
 
 		Path uploadPath = Paths.get(uploadDir);
 
-		if(!Files.exists(uploadPath))
+		if (!Files.exists(uploadPath))
 			Files.createDirectories(uploadPath);
-		
+
+		System.out.println(uploadPath);
+
 		try (InputStream inputStream = multipartFile.getInputStream()) {
 			Path filePath = uploadPath.resolve(fileName);
 			Files.copy(inputStream, filePath, StandardCopyOption.REPLACE_EXISTING);
